% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethod_estMMD.R
\name{summary.estMMD}
\alias{summary.estMMD}
\title{Summary method for the \code{class} \code{"estMMD"}}
\usage{
\method{summary}{estMMD}(object, ...)
}
\arguments{
\item{object}{An object of \code{class} \code{"estMMD"}.}

\item{...}{Additional arguments (not used).}
}
\value{
No return value, called only to print information on the output of \code{"estMMD"}.
}
\description{
Summary method for the \code{class} \code{"estMMD"}
}
