% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethod_regMMD.R
\name{summary.regMMD}
\alias{summary.regMMD}
\title{Summary method for the \code{class} \code{"regMMD"}}
\usage{
\method{summary}{regMMD}(object, ...)
}
\arguments{
\item{object}{An object of \code{class} \code{"regMMD"}.}

\item{...}{Additional arguments (not used).}
}
\value{
No return value, called only to print information on the output of \code{"regMMD"}.
}
\description{
Summary method for the \code{class} \code{"regMMD"}
}
